requirements.txt
# Production E-Commerce Semantic Search RAG System Requirements

# Core Data Processing
pandas==2.0.3
numpy==1.24.3

# Text Processing & NLP
sentence-transformers==2.2.2
nltk==3.8.1

# Vector Search & Indexing
faiss-cpu==1.7.4
# faiss-gpu==1.7.4  # Use for GPU-accelerated search (requires CUDA)

# LLM Integration
google-generativeai==0.3.0

# Configuration Management
pyyaml==6.0

# Logging & Monitoring
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.0
jupyter==1.0.0
ipython==8.14.0

# Utilities
tqdm==4.65.0

# ── Upgrade 1-4: LLM-as-Judge, Routing, Evaluation, Experiment Tracking ──

# Visualisation (notebooks/analysis.ipynb)
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0          # optional: KDE in faithfulness histogram

# Cost / token estimation (already covered by google-generativeai)
# Optional: OpenAI / Anthropic judge backends
# openai>=1.0.0
# anthropic>=0.20.0
